opensky_integration/authenticator/requirements.txt
flask
requests
pystray
pillow
pyinstaller
can